pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "findata"
version = "0.1.0"
description = "Historical financial data management system with support for equities, indices, and multi-asset data"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "FinData Contributors"}
]
keywords = ["finance", "data", "timeseries", "equity", "index", "historical-data"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "yfinance>=0.2.66",
    "pandas>=1.5.0",
    "numpy>=1.20.0",
    "lxml>=4.9.0",
    "html5lib>=1.1",
    "beautifulsoup4>=4.11.0",
    "requests>=2.28.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
dashboard = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/findata"
Documentation = "https://github.com/yourusername/findata/blob/main/CLAUDE.md"
Repository = "https://github.com/yourusername/findata"
Issues = "https://github.com/yourusername/findata/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
findata = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "examples/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
